102-fibonacci.c

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -9,8 +9,8 @@ int main(void)
 {
 	int count = 3;
 
-	int first = 1, second = 2;
-	int next = first + second;
+	long int first = 1, second = 2;
+	long int next = first + second;
 
 	printf("%lu, ", first);
 	printf("%lu, ", second);
